Update train.py
This modification to the YOLOv7 training code allows for validation on two datasets, one from the same domain and one from an opposing domain. This is achieved by running the test function twice and saving the results on the results.txt file. The plots and w&b loggings remain unchanged. By using this modified code, users can evaluate the model's performance on both similar and dissimilar datasets, which can help in identifying any potential biases in the model.